split MapComponent to HOC and MapContainer
